add avro_arrow for conversions boilerplate
